Backend/requirement.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
idna==3.10
pyproj==3.7.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
polyline==2.0.2
paho-mqtt==2.1.0
gql==3.5.2
python-dateutil==2.9.0
requests-toolbelt==2.32.3